# Replace CHM with HashMap in CHMGraphCache