The map in the CHMGraphCache is initialized during class construction and then never updated again. We can safely use a HashMap by assigning it to a final variable in the GraphCache class to ensure safe publication. Then, the `get()` operations should be faster since there is no need to read from `volatile` variables. This is likely a very minor change, but since it is at the bottom of the stack and is used as we traverse the graph for each node that we visit, it seems like a potentially valuable change.

## Current Benchmark Numbers

I ran `mvn clean compile exec:exec@bench -DbenchArgs="nytimes"` two times in a row to get:

```
nytimes-256-angular.hdf5: 289761 base and 9991 query vectors created, dimensions 256
Build M=16 ef=100 in 67.92s with avg degree 31.98 and 0.72 short edges
Uncompressed vectors
  Query (memory) top 100/1 recall 0.7281 in 2.82s after 43,143,500 nodes visited
  Query (disk) top 100/1 recall 0.7281 in 5.96s after 43,143,500 nodes visited
  Query (memory) top 100/2 recall 0.7775 in 6.13s after 75,900,506 nodes visited
  Query (disk) top 100/2 recall 0.7775 in 11.85s after 75,900,506 nodes visited
BinaryQuantization build in 0.06s,
BinaryQuantization encoded 289,761 vectors [13.26 MB] in 0.12s
  Query top 100/1 recall 0.3836 in 2.72s after 54,455,368 nodes visited
  Query top 100/2 recall 0.4993 in 3.24s after 95,710,444 nodes visited
ProductQuantization(64) build in 10.80s,
ProductQuantization(64) encoded 289,761 vectors [22.61 MB] in 2.69s
  Query top 100/1 recall 0.6457 in 2.54s after 44,194,642 nodes visited
  Query top 100/2 recall 0.7638 in 3.97s after 77,578,322 nodes visited
ProductQuantization(32) build in 9.50s,
ProductQuantization(32) encoded 289,761 vectors [13.64 MB] in 2.26s
  Query top 100/1 recall 0.5286 in 1.96s after 46,249,428 nodes visited
  Query top 100/2 recall 0.6695 in 3.19s after 80,703,390 nodes visited

nytimes-256-angular.hdf5: 289761 base and 9991 query vectors created, dimensions 256
Build M=16 ef=100 in 69.08s with avg degree 31.98 and 0.72 short edges
Uncompressed vectors
  Query (memory) top 100/1 recall 0.7271 in 2.61s after 43,040,146 nodes visited
  Query (disk) top 100/1 recall 0.7271 in 5.54s after 43,040,146 nodes visited
  Query (memory) top 100/2 recall 0.7778 in 6.10s after 75,913,800 nodes visited
  Query (disk) top 100/2 recall 0.7778 in 11.70s after 75,913,800 nodes visited
BinaryQuantization build in 0.05s,
BinaryQuantization encoded 289,761 vectors [13.26 MB] in 0.12s
  Query top 100/1 recall 0.3843 in 3.12s after 54,337,928 nodes visited
  Query top 100/2 recall 0.4998 in 4.85s after 95,492,742 nodes visited
ProductQuantization(64) build in 15.86s,
ProductQuantization(64) encoded 289,761 vectors [22.61 MB] in 3.21s
  Query top 100/1 recall 0.6460 in 2.97s after 43,980,682 nodes visited
  Query top 100/2 recall 0.7634 in 4.41s after 77,401,564 nodes visited
ProductQuantization(32) build in 11.15s,
ProductQuantization(32) encoded 289,761 vectors [13.64 MB] in 2.45s
  Query top 100/1 recall 0.5296 in 2.42s after 46,308,822 nodes visited
  Query top 100/2 recall 0.6698 in 3.89s after 80,838,354 nodes visited
```

## With PR

I ran `mvn clean compile exec:exec@bench -DbenchArgs="nytimes"` two times in a row to get:

```
nytimes-256-angular.hdf5: 289761 base and 9991 query vectors created, dimensions 256
Build M=16 ef=100 in 68.41s with avg degree 31.98 and 0.72 short edges
Uncompressed vectors
  Query (memory) top 100/1 recall 0.7288 in 2.76s after 43,148,414 nodes visited
  Query (disk) top 100/1 recall 0.7288 in 5.61s after 43,148,414 nodes visited
  Query (memory) top 100/2 recall 0.7777 in 4.62s after 75,984,238 nodes visited
  Query (disk) top 100/2 recall 0.7777 in 9.17s after 75,984,238 nodes visited
BinaryQuantization build in 0.05s,
BinaryQuantization encoded 289,761 vectors [13.26 MB] in 0.12s
  Query top 100/1 recall 0.3837 in 2.53s after 54,465,400 nodes visited
  Query top 100/2 recall 0.4993 in 2.89s after 95,726,412 nodes visited
ProductQuantization(64) build in 11.57s,
ProductQuantization(64) encoded 289,761 vectors [22.61 MB] in 2.98s
  Query top 100/1 recall 0.6472 in 2.63s after 44,069,632 nodes visited
  Query top 100/2 recall 0.7644 in 4.19s after 77,371,530 nodes visited
ProductQuantization(32) build in 10.44s,
ProductQuantization(32) encoded 289,761 vectors [13.64 MB] in 2.43s
  Query top 100/1 recall 0.5302 in 2.11s after 46,361,400 nodes visited
  Query top 100/2 recall 0.6705 in 3.45s after 80,771,098 nodes visited

nytimes-256-angular.hdf5: 289761 base and 9991 query vectors created, dimensions 256
Build M=16 ef=100 in 73.78s with avg degree 31.98 and 0.72 short edges
Uncompressed vectors
  Query (memory) top 100/1 recall 0.7278 in 2.77s after 43,076,102 nodes visited
  Query (disk) top 100/1 recall 0.7278 in 5.37s after 43,076,102 nodes visited
  Query (memory) top 100/2 recall 0.7781 in 4.48s after 75,968,194 nodes visited
  Query (disk) top 100/2 recall 0.7781 in 8.75s after 75,968,194 nodes visited
BinaryQuantization build in 0.05s,
BinaryQuantization encoded 289,761 vectors [13.26 MB] in 0.12s
  Query top 100/1 recall 0.3833 in 2.41s after 54,300,148 nodes visited
  Query top 100/2 recall 0.4990 in 2.99s after 95,545,502 nodes visited
ProductQuantization(64) build in 10.29s,
ProductQuantization(64) encoded 289,761 vectors [22.61 MB] in 2.49s
  Query top 100/1 recall 0.6473 in 2.50s after 44,135,836 nodes visited
  Query top 100/2 recall 0.7637 in 4.11s after 77,432,240 nodes visited
ProductQuantization(32) build in 9.28s,
ProductQuantization(32) encoded 289,761 vectors [13.64 MB] in 2.11s
  Query top 100/1 recall 0.5304 in 2.31s after 46,319,646 nodes visited
  Query top 100/2 recall 0.6708 in 3.99s after 80,763,738 nodes visited
```

## Additional thoughts

In most of the above results, we're faster with the HM instead of the CHM. I'm not sure how else we want to improve this though.